move types from pages/FAQPage.tsx to types.ts
