+++ frontend/src/types.ts
@@ -1,0 +1,9 @@
+export type FAQ = {
+  question: string;
+  answer: string;
+};
+
+export type FAQData = {
+  headerName: string;
+  faqs: FAQ[];
+};
